Move PreLaunchHacks, add server arg testmod, fix up some issues.
